requirements.txt
aiosmtplib==2.0.2
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
blinker==1.8.2
certifi==2024.7.4
click==8.1.7
deprecation==2.1.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.112.0
fastapi-cli==0.0.5
fastapi-mail==1.4.1
gotrue==2.6.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
packaging==24.1
pluggy==1.5.0
postgrest==0.16.9
psycopg2==2.9.9
pydantic==2.8.2
pydantic-settings==2.5.2
pydantic_core==2.20.1
Pygments==2.18.0
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
realtime==1.0.6
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
storage3==0.7.7
StrEnum==0.4.15
supabase==2.6.0
supafunc==0.5.1
typer==0.12.3
typing_extensions==4.12.2
tzlocal==5.2
uvicorn==0.30.5
uvloop==0.19.0
watchfiles==0.23.0
websockets==12.0
